Refactor out track selection
