+++ src/components/GraphsPanel/types.ts
@@ -1,0 +1,6 @@
+export type OptionType = {
+  label: string
+  value: string
+  dataType: string
+  icon: React.ReactNode
+}
